requirements.txt
streamlit
litellm
python-dotenv
PyMuPDF